Only set up tf if it doesn't exist

+++ proxy/cideploy.sh
@@ -7,10 +7,12 @@ set -x
 cd $(dirname $0)
 
 # Set up Terraform
-wget https://releases.hashicorp.com/terraform/0.7.13/terraform_0.7.13_linux_amd64.zip
-unzip terraform_0.7.13_linux_amd64.zip
-mkdir $(pwd)/bin
-mv terraform bin
+if [ ! -e "bin/terraform" ]; then
+  wget https://releases.hashicorp.com/terraform/0.7.13/terraform_0.7.13_linux_amd64.zip
+  unzip terraform_0.7.13_linux_amd64.zip
+  mkdir $(pwd)/bin
+  mv terraform bin
+fi
 export PATH="$PATH:$(pwd)/bin"
 
 # Remote configs
